By tracing the MINIMAX algorithm to a certain depth -- here we limited the number of available moves to being a small number (2-3) and a depth of 3-4: we printed out statements when each min and max function returned their best value -- and we also printed out when a node was being evaluated. As a result this is an example of a trace that we got MAX NODE VAL: 3 MAX NODE VAL: 2 MIN Best Value: 2 MAX NODE VAL: 4 MAX NODE VAL: 2 MIN Best Value: 2 MAX Best Value: 2 MAX NODE VAL: 3 MAX NODE VAL: 5 MIN Best Value: 3 MAX NODE VAL: 5 MAX NODE VAL: 3 MIN Best Value: 3 MAX Best Value: 3 MIN Best Value: 2 MAX NODE VAL: 5 MAX NODE VAL: 2 MIN Best Value: 2 MAX NODE VAL: 3 MAX NODE VAL: 5 MIN Best Value: 3 MAX Best Value: 3 MAX NODE VAL: 4 MAX NODE VAL: 1 MIN Best Value: 1 MAX NODE VAL: 4 MAX NODE VAL: 0 MIN Best Value: 0 MAX Best Value: 1 MIN Best Value: 1 AB Best Value: 2 (1, 0)
Len Available Actions: 2
Depth = 4
AB Pruning = OFF

you can draw out the entire tree using this method.

Instead of using an actual evaluation function we used a random number generator for ease of use -- when we used this it can be seen that our algorithm is working as expected. Therefore it could be that our evaluation function is not working correctly

Therefore we need to judge whether the evaluation function is working with the undo moves (are we evaluating moves at the base state instead of at the state of the terminal/cuttoff nodes)