Need to add more features that relate to move geneation and also board evaluation
The greedy algorithm that we currently have is very dependent on how accurate our evaluation is of which moves are considered to be "the best to make at that point in time" -- we use this as an approximation to guide our search and to limit the large branching factor of this game.
    ---- Features to implement:

Need to work out various optimisations relating to algorithmic complexity reduction -- i.e. linear to constant and quadratic to linear run times. Any reduction in this sort of thing inside the exponential recursion will produce a great increase in performance which is neccessary for higher ply searches. This would also mean we are able to open up the window of "favourable moves"

Also still need to consider a way to speed up how we update the board configuration -- this is a major limiting factor of our run time in both minimax and also mcts... we need to be able to simulate a board game within a reasonable amount of time such that we can generate many nodes in the game tree.

Need to implement this greedy nature into the transposition table negamax