trying to work on optimisation of the board game -- i think that most of the bottleneck is in the backend of the game -- therefore I am thinking I might need to rewrite the whole back end of the board game to be more efficient
What I am thinking is that we switch from this one board class into a more OOP design that utilises a class for the piece, and dictionaries that store the pieces that are currently on the board, instead of lists. This would mean that it would be much faster to access. Therefore we should see better improvement in the run time of the adversarial search...

Thinking of rewriting the update_board and undo_move functions
-- if we return a list of pieces that was eliminated when a move was made, if we call undo move straight after making this move ,then we can use those eliminated piece list to restore the pieces on the board, I think this would be a much faster method of undo-ing board moves instead of using seperate stack to store all eliminated pieces throughout the board game. This is just added complexity and probably is impacting on our running time the most.